+++ src/components/FAQ.jsx
@@ -1,0 +1,55 @@
+"use client"
+import { Accordion, Span, Stack, Text } from "@chakra-ui/react"
+import { useState } from "react"
+
+export function FAQ() {
+    const [value, setValue] = useState(["item1"])
+
+    const items = [
+        {
+            value: "item1",
+            title: "Q: What makes XYZ Company's products so unique?",
+            text: "A: At XYZ Company, we pride ourselves on innovation and cutting-edge technology. Our products are crafted with a perfect blend of creativity and export functionality, setting them apart from the competition. We constantly strive to push the boundaries of what's possible to deliver a truly exceptional user experience."
+        },
+        {
+            value: "item2",
+            title: "Q: How does XYZ Company contribute to sustainability and environmental conservation?",
+            text: "A: XYZ Company is committed to environmental responsibility. We implement eco-friendly practices throughout our production process, from sourcing materials to manufacturing. Our dedication to sustainability extends beyond our products, as we actively engage in initiatives to reduce our carbon footprint and promote a greener future."
+        },
+        {
+            value: "item3",
+            title: "Q: What kind of support can customers expect from XYZ Company's customer service team?",
+            text: "A: At XYZ Company, customer satisfaction is our top priority. Our dedicated support team is available 24/7 to address any inquiries or concerns you may have. Whether you need assistance with product usage, troubleshooting, or general information, our knowledgeable and friendly customer service representatives are always ready to provide prompt and effective support."
+        },
+        {
+            value: "item4",
+            title: "Q: How does XYZ Company stay ahead of industry trends?",
+            text: "A: XYZ Company stays at the forefront of industry trends through continuous research and development. We invest heavily in staying ahead of the curve, monitoring market dynamics, and actively seeking feedback from our customers. By embracing innovation and staying agile, we ensure that our products and services evolve to meet the ever-changing needs of our dynamic customer base."
+        },
+        {
+            value: "item5",
+            title: "Q: Can you tell us about XYZ Company's commitment to community involvement?",
+            text: "A: XYZ Company is deeply committed to giving back to the communities we serve. We actively engage in philanthropic initiatives, supporting local charities and community projects. Additionally, we encourage our employees to participate in volunteer programs, fostering a culture of social responsibility. By being actively involved in the community, we aim to make a positive impact beyond the products and services we offer."
+        },
+
+    ]
+
+    return (
+        <Stack gap="1" className="faq-container">
+            <Text fontWeight="medium" className="faq-title">FAQ (คำถามที่พบบ่อย)</Text>
+            <Accordion.Root value={value} onValueChange={(e) => setValue(e.value)} className="faq-list">
+                {items.map((item, index) => (
+                    <Accordion.Item key={index} value={item.value} className="faq-item">
+                        <Accordion.ItemTrigger>
+                            <Span flex="1" className="arrow">{item.title}</Span>
+                            <Accordion.ItemIndicator />
+                        </Accordion.ItemTrigger>
+                        <Accordion.ItemContent>
+                            <Accordion.ItemBody className="faq-answer">{item.text}</Accordion.ItemBody>
+                        </Accordion.ItemContent>
+                    </Accordion.Item>
+                ))}
+            </Accordion.Root>
+        </Stack>
+    );
+}
