+++ src/components/Header.jsx
@@ -1,0 +1,23 @@
+export function Header() {
+    return (
+        <header>
+            <nav>
+                <div className="logo">🚀 XYZ CORP</div>
+                <ul className="nav-list">
+                    <li>
+                        <a href="#">Home</a>
+                    </li>
+                    <li>
+                        <a href="#">About</a>
+                    </li>
+                    <li>
+                        <a href="#">Services</a>
+                    </li>
+                    <li>
+                        <a href="#">Contact</a>
+                    </li>
+                </ul>
+            </nav>
+        </header>
+    );
+}
